requirements.txt
fastapi
uvicorn
python-multipart
Pillow
numpy
scikit-learn
aiofiles
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
